add useImport
