+++ src/pages/hooks/useImport/use-import.types.ts
@@ -1,0 +1,14 @@
+import { Account, AccountSource } from "../../../types";
+
+export interface UseImportProps {
+  accounts: Account[];
+  updateAccounts: (newAccounts: Account[]) => void;
+}
+
+export interface ImportState {
+  startImport: (
+    csvData: string | ArrayBuffer | null,
+    source: AccountSource | null
+  ) => void;
+  onResetAccounts: () => void;
+}
